添加shader 	new file:   unity3d/ShaderProject/Assets/Materials/M4.mat 	new file:   unity3d/ShaderProject/Assets/Materials/M4.mat.meta 	new file:   unity3d/ShaderProject/Assets/Materials/M5.mat 	new file:   unity3d/ShaderProject/Assets/Materials/M5.mat.meta 	new file:   unity3d/ShaderProject/Assets/Materials/M6.mat 	new file:   unity3d/ShaderProject/Assets/Materials/M6.mat.meta 	new file:   unity3d/ShaderProject/Assets/Materials/M7.mat 	new file:   unity3d/ShaderProject/Assets/Materials/M7.mat.meta 	new file:   unity3d/ShaderProject/Assets/Materials/M8.mat 	new file:   unity3d/ShaderProject/Assets/Materials/M8.mat.meta 	new file:   unity3d/ShaderProject/Assets/Materials/M9.mat 	new file:   unity3d/ShaderProject/Assets/Materials/M9.mat.meta 	modified:   unity3d/ShaderProject/Assets/Scene/Main.unity 	modified:   unity3d/ShaderProject/Assets/Shaders/03-Shader.shader 	new file:   unity3d/ShaderProject/Assets/Shaders/04-Shader.shader 	new file:   unity3d/ShaderProject/Assets/Shaders/04-Shader.shader.meta 	new file:   unity3d/ShaderProject/Assets/Shaders/05-Shader.shader 	new file:   unity3d/ShaderProject/Assets/Shaders/05-Shader.shader.meta 	new file:   unity3d/ShaderProject/Assets/Shaders/06-Diffuse Fragment HalfLambert.shader 	new file:   unity3d/ShaderProject/Assets/Shaders/06-Diffuse Fragment HalfLambert.shader.meta 	new file:   unity3d/ShaderProject/Assets/Shaders/07-Specular Vertex.shader 	new file:   unity3d/ShaderProject/Assets/Shaders/07-Specular Vertex.shader.meta 	new file:   unity3d/ShaderProject/Assets/Shaders/08-Specular Fragment.shader 	new file:   unity3d/ShaderProject/Assets/Shaders/08-Specular Fragment.shader.meta 	new file:   unity3d/ShaderProject/Assets/Shaders/09-Specular Fragment BlinnPhong.shader 	new file:   unity3d/ShaderProject/Assets/Shaders/09-Specular Fragment BlinnPhong.shader.meta

+++ unity3d/ShaderProject/Assets/Shaders/03-Shader.shader
@@ -19,8 +19,8 @@
 			};
 
 			struct v2f {
-				fixed4 position : SV_POSITION;
-				float3 temp : COLOR0;
+				float4 position : SV_POSITION;
+				fixed3 temp : COLOR0;
 			};
 
 			v2f vert(a2v v) {
